==> parks/13_create_park_appendices.R <==
## This script creates several table-based appendices documenting parks which were excluded
## from analyses, dropped for coverage issues, etc.

##### parks excluded from all analyses #####
all_parks_metadata <- readRDS(file.path(here(), "data-intermediate/parks/all-parks-metadata.RDS"))

dropped_entirely <- all_parks_metadata %>%
  filter(!zone_name %in% park_metadata$zone_name) %>%
  select(
    system = system_label,
    unit_name = unit_label
  ) %>%
  mutate(reason = "Missing >50% of all weekly LBS records.")

dropped_pre_LBS <- data.frame(
  system = c("DNR State", "DNR State", "Metro Regional", "DNR State"),
  unit_name = c(
    "Garden Island (State Recreation Area)",
    "Saint Croix Islands (State Recreation Area)",
    "Sam Brown Memorial (State Wayside)",
    "Lake Minnetonka Islands, Three Rivers"
  ),
  reason = c("Island", "Island", "Less than 10,000 square kilometers", "Island")
)

dropped_bike_only <- park_metadata %>%
  filter(str_detect(unit_label, "Franz|Latsch|Greenleaf|Kilen|Collinwood")) %>%
  select(
    system = system_label,
    unit_name = unit_label
  ) %>%
  mutate(reason = "No LBS bike data was returned due to low data coverage.")

##### parks for which no bike imputation was performed #####

load(file.path(here(), "data-intermediate/parks/filled-imputed-volume.rda"))

no_bike_imputation <- no_bike_imputation %>%
  left_join(park_metadata) %>%
  transmute(
    system = system_label,
    unit_name = unit_label
  )

##### parks where we processed geography differently #####

### manually document based on code in 02_ processing steps ###
geography_exceptions <- data.frame(
  unit_label = c(
    "Above the Falls, MPRB",
    "Afton",
    "Alexander Ramsey",
    "Appleton OHV (Eligible)",
    "Banning",
    "Bear Head Lake",
    "Bend in the River (GRPC)",
    "Big Stone Lake",
    "Bluffs Traverse",
    "Cascade River",
    "Central Mississippi Riverfront, MPRB",
    "Cohasset Tioga",
    "Collinwood",
    "Fish Lake (Eligible)",
    "Fort Ridgely",
    "Fort Snelling",
    "Frontenac",
    "Garvin",
    "Glendalough",
    "Gooseberry Falls",
    "Grand Marais Recreation Area (Eligible)",
    "Grand Portage",
    "Granite Falls Memorial",
    "Hayes Lake",
    "He Mni Can - Barn Bluff",
    "Hyland-Bush-Anderson Lakes, Bloomington",
    "Hyland-Bush-Anderson Lakes, Three Rivers",
    "Irving and John Anderson",
    "Itasca",
    "Jay C. Hormel Nature Center",
    "Keller, Ramsey County",
    "Kensington Rune Stone",
    "Kilen Woods",
    "Kodonce River (State Wayside)",
    "Lac qui Parle (DNR)",
    "Lake Bronson",
    "Lake Byllesby, Dakota County",
    "Lake Carlos",
    "Lake Elmo, Washington County",
    "Lake George, Anoka County",
    "Lake Sarah, Three Rivers",
    "Lake Shetek",
    "Lake Vermilion-Soudan Underground Mine",
    "Lake Waconia, Carver County",
    "Lebanon Hills, Dakota County",
    "Lilydale-Harriet Island, Saint Paul",
    "Long Lake, Ramsey County",
    "Maplewood",
    "McCarthy Beach",
    "Miesville Ravine, Dakota County",
    "Minneapolis Chain of Lakes, MPRB",
    "Minnehaha, MPRB",
    "Mississippi Gateway, Three Rivers",
    "Mississippi Gorge, MPRB",
    "Mississippi Gorge, Saint Paul",
    "Noerenberg Gardens, Three Rivers",
    "Nokomis-Hiawatha, MPRB",
    "North Mississippi, MPRB",
    "North Mississippi, Three Rivers",
    "Phalen, Saint Paul",
    "Pine Point, Washington County",
    "Pine Valley (Eligible)",
    "Plum Creek",
    "Ray Berglund (State Wayside)",
    "Rice Creek Chain of Lakes, Anoka County",
    "Rice Lake",
    "Robbins Island",
    "Robert Ney Memorial",
    "Rockville",
    "Root River (Eligible)",
    "Rum River Central, Anoka County",
    "St. Croix Bluffs, Washington County",
    "Saint Croix Boom Site (State Wayside)",
    "Sandstone Robinson",
    "Spirit Mountain",
    "Split Rock Lighthouse",
    "Spring Lake, Dakota County",
    "Spring Lake, Scott County",
    "Stanley Eddy",
    "Temperance River",
    "Theodore Wirth, MPRB",
    "Twin Lakes",
    "Upper Sioux Agency",
    "Whitetail Woods, Dakota County"
  ),
  exception = c(
    "Reduced road buffer",
    "Reduced road buffer to 10m",
    "Reduced road buffer to 5m",
    "Reduced road buffer to 25m",
    "Based on feedback, manually included additional parking lot",
    "Expanded water buffer to create consistent shoreline",
    "Reduced road buffer to 20m",
    "Reduced road buffer to 20m, expanded water buffer to create consistent shoreline, manually added overlook area",
    "Manually included additional parking",
    "Based on feedback, included section of road previously removed",
    "Reduced buffer on all non-highway roads",
    "Reduced road buffer, reduced water buffer to create consistent shoreline",
    "Reduced buffer on city roads to 15m",
    "Original shapefile was misaligned: edited polygon to reflect published maps of the park, reduced road buffer to 10m. Further refinement requires revised shapefile.",
    "Based on feedback, included historic Fort Ridgely inholding",
    "Reduced road buffer to 20m",
    "Based on feedback, manually included additional parking lot",
    "Reduced road buffer to 10m",
    "Based on feedback, manually included Molly Stark boat launch",
    "Based on feedback, reduced road buffer to 15m to retain Gitchi Gami state trail",
    "Reduced road buffer to 10m",
    "Based on feedback, manually included additional parking",
    "Reduced road buffer to 20m",
    "Based on Feedback, included section of road previously removed",
    "Reduced select road buffers to 5m",
    "Reduced select road buffers to 5m",
    "Reduced select road buffers to 5m",
    "Reduced road buffers to 20m",
    "Based on feedback, included SNA",
    "Removed private property (Gerard Academy, Austin Country Club)",
    "Reduced road buffers to 5m",
    "Increased water buffer to create consistent shoreline",
    "Reduced road buffers near bridge to 10m",
    "Based on feedback, manually included additional parking",
    "Based on feedback, manually included park office which lies outside of park boundary",
    "Reduced select road buffers to 5m",
    "Based on feedback, expanded park boundary 50ft east of Harry Ave",
    "Based on feedback, manually included lazy river parking lot",
    "Increased water buffer to create consistent shoreline",
    "Reduced road buffers to 5m",
    "Reduced road buffers to 15m",
    "Based on feedback, remove previously included cabins",
    "Based on feedback, manually included additional parking",
    "Based on feedback, manually removed access to DNR managed boat launch. Further refinement requires revised shapefile.",
    "Based on feedback, removed Camp Butwin inholding",
    "Reduced road buffers to 15m",
    "Reduced road buffers to 15m",
    "Based on feedback, removed previously included cabins",
    "Based on feedback, manually included additional parking",
    "Based on feedback, manually included portion of park in Goodhue County. Further refinement requires revised shapefile.",
    "Created road-inclusive and road-exclusive polygons. Reduced road buffers to 15m in road-exlusive polygon.",
    "Created road-inclusive and road-exclusive polygons. Reduced road buffers to 15m in road-exlusive polygon.",
    "Reduced road buffer to 15m",
    "Reduced road buffer to 5m",
    "Reduced road buffer to 5m",
    "Reduced road buffer to 20m",
    "Created road-inclusive and road-exclusive polygons. Reduced road buffers to 20m in road-exclusive polygon.",
    "Reduced road buffer to 15m",
    "Reduced road buffer to 15m",
    "Created road-inclusive and road-exclusive polygons. Reduced road buffers to 5m in road-exclusive polygon.",
    "Reduced road buffer to 15m",
    "Reduced road buffer to 15m",
    "Reduced road buffer to 15",
    "Reduced road buffer to 3m",
    "Reduced road buffer to 5-20m",
    "Based on feedback, manually included park entrance and filled in gap in original shapefile",
    "Reduced road buffer to 5m",
    "Reduced road buffer to 5m",
    "Reduced road buffer to 15m",
    "Increased water buffer to create consistent shoreline",
    "Increased water buffer to create consistent shoreline",
    "Based on feedback, included additional land",
    "Reduced road buffer to 3m",
    "Reduced road buffer to 15m",
    "Reduced road buffer to 15m",
    "Based on feedback, excluded lighthouse, included Gitchi Gami and scenic outlook",
    "Based on feedback, removed previously included road segments",
    "Based on feedback, manually added two parking lots",
    "Original shapefile was misaligned: edited polygon alignment to reflect published maps of the park. Reduced road buffer to 20m. Further refinement requires revised shapefile.",
    "Based on feedback, manually included additional parking",
    "Reduced road buffer to 20m",
    "Increased water buffer to create consistent shoreline",
    "Based on feedback, manually included historical site",
    "Based on feedback, manually included trailhead"
  )
) %>%
  left_join(park_metadata %>% select(unit_label, system_label)) %>%
  mutate(system_label = ifelse(str_detect(unit_label, "Parle"), "DNR State", system_label))

## check:
# geography_exceptions %>% filter(!unit_label %in% park_metadata$unit_label)

##### parks with unresolved geography issues #####

unresolved_geography <- data.frame(
  unit_label = c(
    "Above the Falls, MPRB",
    "Bluffs Traverse",
    "Central Mississippi Riverfront, MPRB",
    "Cohasset Tioga",
    "Cuyuna Country (State Recreation Area)",
    "Hyland-Bush-Anderson Lakes, both units",
    "Fish Lake (Eligible)",
    "Frazee (Eligible)",
    "Gooseberry Falls",
    "Grand Marais Recreation Area (Eligible)",
    "Grand Portage",
    "Kodonce River (State Wayside)",
    "Lake Brophy",
    "Mississippi Gorge, MPRB",
    "Mississippi River (GRPC)",
    "North Mississippi, Three Rivers",
    "Robbins Island",
    "Sawtooth Bluff (Eligible)",
    "Split Rock Lighthouse",
    "Stanley Eddy",
    "Temperance River",
    "Tettegouche"
  ),
  concern = c(
    "Narrow",
    "Parking",
    "Narrow",
    "Complicated geography",
    "Complicated geography",
    "Original shapefile incorrectly assigned parcels to implementing agencies",
    "Misaligned shapefile",
    "Roads could not be removed",
    "North Shore",
    "Roads could not be removed",
    "North Shore",
    "North Shore",
    "Water boundary may include private waterfront and/or exclude public waterfront",
    "Narrow",
    "Narrow",
    "Narrow",
    "The portion of the park on County Hwy 41 (Rau park unit) is challenging to separate from the highway itself.",
    "Parking",
    "North Shore",
    "Misaligned shapefile",
    "North Shore",
    "North Shore"
  )
) %>%
  left_join(park_metadata %>% select(unit_label, system_label)) %>%
  mutate(system_label = ifelse(str_detect(unit_label, "Hyland"), "Metro Regional", system_label))

##### save #####
save(dropped_entirely, dropped_pre_LBS, dropped_bike_only,
  no_bike_imputation, geography_exceptions, unresolved_geography,
  file = file.path(here(), "data-intermediate/documentation/park-appendices.rda")
)
